+++ lib/generators/heap_periscope_agent/config_generator.rb
@@ -2,7 +2,7 @@ require 'rails/generators/base'
 
 module HeapPeriscopeAgent
   module Generators
-    class InstallGenerator < Rails::Generators::Base
+    class ConfigGenerator < Rails::Generators::Base
       source_root File.expand_path("templates", __dir__)
 
       def copy_initializer_file